Removed BASENAME_ONLY from linking.
The use of BASENAME_ONLY is discouraged by LArSoft.
One effect of it is possible library name conflicts, that indeed happens due to our love for cut&paste.
An example in our repository is AnalysisTree module: each of the experiments has it.
This makes it impossible to have multiple experiments in the same repository (although other reason might still preventthat, as conflicting dependencies).

Note that the correct specification for ambuguous module is, e.g.:

    module_type: "lar1ndcode/AnalisysTree/AnalysisTree"