add data-driven adc simulation
Initial version of a producer module which takes simulated events and applies
the (linearly, or perfectly calibrated) test stand data, and removes bad codes
(currently LSBs are 000000, 000001, or 111111), and stores that as a new data
product (which we could run hit finding on) and also outputs user-friendly
before-and-after ROOT TH2Ses into a file defined with TFileService for each
plane, for quick analysis.

The test stand data into a mapping from "ideal" ADC based on input voltage
to the distribution of measured output codes, and the "simulation" is done
by just sampling from the appropriate distribution. For now I'm using data
from the chips considered good based on data loss fractions, distributed
randomly around the detector.

N.B. It is recommended to adjust the baselines to avoid sticky regions
(e.g. the induction plane default in 2048), and disable waveform noise to
avoid double-counting.